Reverted to old method
Fixed the pagination problem. Thought I could use the filtrer service from Tableau but it does not work it clients have
",", "&" or any special character usually used in URL QueryString.
The solution is then to get all the projects and iterate till finding the right one or sending an
error if not found after going through all the pages.